requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
numpy==1.15.2
numpydoc==0.7.0
Pillow==4.2.1
requests==2.18.4
tensorboard==1.10.0
tensorflow==1.10.1

